% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_gumbel.R
\name{scale_x_gumbelRP}
\alias{scale_x_gumbelRP}
\alias{scale_y_gumbelRP}
\alias{scale_x_gumbelAEP}
\alias{scale_y_gumbelAEP}
\title{Gumbel-transformed axes}
\usage{
scale_x_gumbelRP(
  ...,
  minor_breaks = unlist(lapply(10^(-10:10), function(x) x * 1:10))
)

scale_y_gumbelRP(
  ...,
  minor_breaks = unlist(lapply(10^(-10:10), function(x) x * 1:10))
)

scale_x_gumbelAEP(
  ...,
  minor_breaks = unlist(lapply(10^(-10:10), function(x) x * 1:10))
)

scale_y_gumbelAEP(
  ...,
  minor_breaks = unlist(lapply(10^(-10:10), function(x) x * 1:10))
)
}
\arguments{
\item{...}{Arguments to pass to `scale_x_continuous` or `scale_y_continuous`
from ggplot2.}

\item{minor_breaks}{Minor grid lines; passed to `scale_x_continuous` or
`scale_y_continuous`.}
}
\value{
The same output as `scale_x_continuous` or `scale_y_continuous`,
but with the appropriate gumbel spacing.
}
\description{
Transforms return period or AEP on either the x or y axis to be spaced
according to a reduced Gumbel variate, meant to be added as a layer to
a ggplot2 graphic.
}
