% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_gumbel.R
\docType{data}
\name{gumbelRP_trans}
\alias{gumbelRP_trans}
\title{Gumbel transformations used for ggplot2 scales}
\format{
An object of class \code{trans} of length 7.
}
\usage{
gumbelRP_trans
}
\description{
Gumbel transformations used for ggplot2 scales
}
\keyword{datasets}
