% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pppoints}
\alias{pppoints}
\title{Probability-Probability Points}
\usage{
pppoints(..., data, ngrid, a = -0.5, col_prefix = "prob")
}
\arguments{
\item{...}{Numeric vectors (column names when `data` is specified), or
distplyr distributions.}

\item{data}{Optional data frame specifying where to take the numeric vectors
from.}

\item{ngrid}{Number of uniform scores to evaluate the quantiles at.}

\item{a}{Uniform score adjustment value, passed to `uscore()` when
determining the uniform grid to evaluate the quantiles on.}

\item{col_prefix}{Prefix for output column names.}
}
\value{
A tibble of paired quantiles, each row corresponding to a
uniform score (non-exceedance probability).
}
\description{
Evaluates non-exceedance probabilities (cumulative distribution function)
on a grid of uniform scores.
}
\note{
Code is a little hacky right now.
}
\examples{
set.seed(1)
qqpoints(rnorm(10), rnorm(10))
qqpoints(rnorm(10), dst_norm(0, 1))
qqpoints(distionary::dst_t(5), distionary::dst_norm(0, 1), ngrid = 20)
}
