% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R, R/qq.R
\name{pppairs}
\alias{pppairs}
\alias{qqpairs}
\title{QQ and PP pairs}
\usage{
pppairs(x, y, data, ngrid, a = -0.5, col_prefix = "prob")

qqpairs(x, y, data, ngrid, a = -0.5, col_prefix = "quantile")
}
\arguments{
\item{x, y}{Numeric vectors or (univariate) probability distributions (of
class `"dst"`); when `data` is specified, these objects will be searched
for in the data first.}

\item{data}{Optional data frame specifying where to look for the `x` and `y`
objects.}

\item{ngrid}{Grid size to evaluate the pairs. Optional if `x` or `y` is
numeric or a `"finite"` distribution, in which case the grid is
made up of the discrete points. In the case of `pppairs()`, the grid
of uniform scores is associated with `x`.}

\item{a}{Uniform score adjustment value when determining non-exceedance
probabilities associated with `x` or `y`, if numeric; or, the grid
of uniform scores if `ngrid` is supplied. Passed to `uscore()`.}

\item{col_prefix}{Prefix for output column names.}
}
\value{
A tibble of paired quantiles (`qqpairs()`) or
non-exceedance probabilities (`pppairs()`), the first column associated with
`x`, and the second with `y`.
}
\description{
Pairs quantiles (`qqpoints()`) and non-exceedance probabilities
(`pppairs()`)
on a grid of uniform scores, for the production of QQ and PP plots.
}
\examples{
set.seed(1)
qqpairs(rnorm(10), rnorm(10))
qqpairs(rnorm(10), distionary::dst_norm(0, 1))
qqpairs(distionary::dst_t(5), distionary::dst_norm(0, 1), ngrid = 20)
pppairs(rnorm(10), rnorm(10))
pppairs(rnorm(10), distionary::dst_norm(0, 1))
pppairs(distionary::dst_t(5), distionary::dst_norm(0, 1), ngrid = 20)
}
