% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{uscore}
\alias{uscore}
\alias{nscore}
\alias{rpscore}
\title{Rank-based Scores}
\usage{
uscore(x, pos = "Hazen", na.rm = FALSE)

nscore(x, pos = "Hazen", na.rm = FALSE)

rpscore(x, pos = "Hazen", na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{pos}{Positional adjustment for uniform scores. See Details.
Can be a single numeric, or could be named after one of the
proponents behind a choice of the numeric: "Weibull", "Beard",
"Gringorten", or "Hazen".}

\item{na.rm}{Logical indicating whether `NA` and `NaN` values should be
removed from the output.}
}
\value{
Vector of uniform scores.
}
\description{
Converts a numeric vector to its rank-based scores. For `uscore()`
(uniform scores), values become roughly equally spaced between 0 and 1,
keeping their order. `nscore()` calculates normal scores by
spacing the uniform scores along a standard
normal distribution; `rpscore()` calculates empirical return periods
by spacing the uniform scores `u` by `1 / (1 - u)`.
}
\details{
Uniform scores are calculated by `(rank+a)/(n+1+2*a)`,
where `rank` is the ranked `x` values, and `a` is the positional
adjustment `pos`. Alternatively, could be named after an individual
associated with a choice of `a`:

- Weibull (1939) proposed `a = 0`.
- Beard (1943) proposed `a = -0.31`.
- Gringorten (1963) proposed `a = -0.44`.
- Hazen (1914) proposed `a = -0.5`.
}
\examples{
x <- c(0.3, 0.56, NA, 0.1, -12)
uscore(x)
uscore(x, pos = "Gringorten")
nscore(x, pos = -0.4)
rpscore(x)

}
\author{
Thanks to Dr. Harry Joe for providing a starting framework for
the `uscore()` function.
}
