% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{uscore}
\alias{uscore}
\alias{nscore}
\alias{rpscore}
\title{Rank-based Scores}
\usage{
uscore(x, a = -0.5, na.rm = FALSE, ...)

nscore(x, a = -0.5, na.rm = FALSE, ...)

rpscore(x, a = -0.5, na.rm = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{a}{Single numeric; adjustment value for uniform scores. See Details.}

\item{na.rm}{Logical indicating whether `NA` and `NaN` values should be
removed from the output.}

\item{...}{Other arguments to pass to the `rank()` function.}
}
\value{
Vector of uniform scores.
}
\description{
Converts a numeric vector to its rank-based scores. For `uscore()`
(uniform scores), values become roughly equally spaced between 0 and 1,
keeping their order. `nscore()` calculates normal scores by
spacing the uniform scores along a standard
normal distribution; `rpscore()` calculates empirical return periods
by spacing the uniform scores `u` by `1 / (1 - u)`.
}
\details{
Uniform scores are calculated by `(rank+a)/(n+1+2*a)`,
where `rank` is the ranked `x` values.
}
\examples{
x <- c(0.3, 0.56, NA, 0.1, -12)
uscore(x)
nscore(x)
rpscore(x)

}
\author{
Thanks to Dr. Harry Joe for providing a starting framework for
the `uscore()` function.
}
